added sub packages for cassandra and mongo
